+++ packages/codemod/lib/utils/index.ts
@@ -1,5 +1,5 @@
 export * from './addMissingImports';
+export * from './filterOutImports';
 export * from './generateLevelTokens';
 export * from './expectTransformFactory';
-export * from './typeProps';
 export * from './varToMemberExpression';
