+++ packages/codemod/lib/v2/index.ts
@@ -1,8 +1,11 @@
 import {Transform} from 'jscodeshift';
-import migrateOldTokens from './migrateOldTokens';
+import migrateColorTokens from './migrateColorTokens';
+import migrateDepthTokens from './migrateDepthTokens';
+import migrateOtherTokens from './migrateOtherTokens';
+import migrateTypeTokens from './migrateTypeTokens';
 
 const transform: Transform = (file, api, options) => {
-  const fixes = [migrateOldTokens];
+  const fixes = [migrateColorTokens, migrateDepthTokens, migrateTypeTokens, migrateOtherTokens];
   return fixes.reduce((source, fix) => fix({...file, source}, api, options) as string, file.source);
 };
