+++ packages/codemod/lib/v2/mapping/typeProps.ts
